requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
shap>=0.42.0
lifelines>=0.27.0
PyYAML>=6.0
jupyter>=1.0.0
pytest>=7.4.0
tqdm>=4.65.0
optuna>=3.3.0
